Move objects again
