+++ src/DeepLinkResources/Resource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Packback\Lti1p3\DeepLinkResource;
+namespace Packback\Lti1p3\DeepLinkResources;
 
 use Packback\Lti1p3\LtiConstants;
 use Packback\Lti1p3\LtiLineitem;
